Refactor: Move models into their own file
